move dynamic stuff to the pkg/dynamic package
