+++ pkg/dynamic/field.go
@@ -1,0 +1,8 @@
+package dynamic
+
+import "reflect"
+
+func HasField(rs reflect.Value, fieldName string) (field reflect.Value, ok bool) {
+	field = rs.FieldByName(fieldName)
+	return field, field.IsValid()
+}
